+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 22.02+ (???)
 ------------------------------------------------------------------------
+- Feature: [#1327] Readd the game intro (use commandline switch --intro to enable)
 - Fix: [#1320] Inability to mark scenario as complete.
 - Fix: [#1325] Crash when saving second loaded scenario of a playthrough.